requirements-base.txt
# Base dependencies used across multiple scripts
pandas>=2.0.0
numpy>=1.26.0
pytz>=2023.3
matplotlib>=3.7.0
pyyaml>=6.0
requests>=2.32.0
finnhub-python>=1.4.0
